+++ src/fireball/LOADFDATA/readdata_2c.f90
@@ -1,11 +1,11 @@
 subroutine readdata_2c (interaction, iounit, num_nonzero, numz, zmax, itype, in1, in2)
-  use iso_c_binding
+  use, intrinsic :: iso_fortran_env, only: double => real64
   use M_fdata, only: ME2c_max, nfofx
   implicit none
-  integer(c_long), intent (in) :: in1, in2, interaction, iounit, itype, num_nonzero, numz
-  real(c_double), intent (in) :: zmax
-  integer(c_long) :: ipoint, integral
-  real(c_double), dimension (ME2c_max, nfofx) :: xintegral_2c
+  integer, intent (in) :: in1, in2, interaction, iounit, itype, num_nonzero, numz
+  real(double), intent (in) :: zmax
+  integer :: ipoint, integral
+  real(double), dimension (ME2c_max, nfofx) :: xintegral_2c
   !if (interaction .ne. 8) then
   xintegral_2c = 0.0d0
   do ipoint = 1, numz
